simplifying OG image generation and removing edge function dependency
