+++ _api/og/postimage/[id].tsx
@@ -1,54 +1,41 @@
-// @ts-nocheck
-
+import { Readable } from 'node:stream';
+// @ts-ignore
+import type { ReadableStream } from 'node:stream/web';
 import React from 'react';
 
-import type { VercelRequest } from '@vercel/node';
+import type { VercelRequest, VercelResponse } from '@vercel/node';
 import { ImageResponse } from '@vercel/og';
+
+import { getPostInfo } from '../getPostInfo.ts';
 
 import { OgPostImage } from './OgPostImage';
 
-export const config = {
-  runtime: 'edge',
-};
-
-export const edge = true;
-
-export default async function handler(req: VercelRequest) {
+export default async function handler(req: VercelRequest, res: VercelResponse) {
   try {
-    const originalUrl = new URL(req.url as string);
+    let urlPath = req.url as string;
+    if (!urlPath.startsWith('http')) {
+      urlPath = `https://www.fake.app${urlPath}`;
+    }
+    const originalUrl = new URL(urlPath);
 
     const parts = originalUrl.pathname.split('/');
     const id = parts[parts.length - 1] ?? 'IDK';
 
-    const url = new URL(
-      'https://' +
-        originalUrl.hostname +
-        '/api/og/postinfo/' +
-        encodeURIComponent(id)
-    );
+    const post = await getPostInfo(id);
+    if (!post?.profile) {
+      return res.status(404).send({
+        message: 'profile not found',
+      });
+    }
 
-    const res = await fetch(url.toString());
-
-    const post: {
-      description: string;
-      profile: {
-        name: string;
-        avatar: string | undefined;
-        links: number;
-        reputation_score: {
-          total_score: number;
-        };
-      };
-    } = await res.json();
-
-    return new ImageResponse(
+    const ir = new ImageResponse(
       (
         <OgPostImage
           description={post.description}
           name={post.profile.name}
           avatar={post.profile.avatar}
           links={post.profile.links}
-          rep={post.profile.reputation_score.total_score}
+          rep={post.profile.reputation_score?.total_score ?? 0}
           scale={2}
         />
       ),
@@ -57,10 +44,11 @@ export default async function handler(req: VercelRequest) {
         height: 630,
       }
     );
+
+    // @ts-ignore
+    Readable.fromWeb(ir.body as ReadableStream<any>).pipe(res);
   } catch (e: any) {
     console.error(`${e.message}`);
-    return new Response(`Failed to generate the image`, {
-      status: 500,
-    });
+    res.status(500).send({ message: 'Failed to generate the image' });
   }
 }
